Added League page, Search Bar, & more History
Added a fourth page called League where I add a search-bar that I want to use with the API key I added in the last push. The plan is to have it pull up NBA players & compare them to Celtics Players eventually, but for now I just want to extract the data from the console. I also add more to the team history page